Return default LayoutRect value When the Layout tree ancestor chain is not available.
https://bugs.webkit.org/show_bug.cgi?id=279419

Reviewed by NOBODY (OOPS!).

When the layout tree ancestor chain is not available, for example, in the case that
the renderer is not yet attached to the tree, We should avoid to accessing the layout
tree ancestor chain.
With the test case specified in cssbox-content-box-002.html, we shouldn't attempt to
work out the padding value when the renderer is not attached to the tree, which causes
the assert failure.

* LayoutTests/TestExpectations:
* Source/WebCore/rendering/RenderBox.cpp:
(WebCore::RenderBox::referenceBoxRect const):