REGRESSION(285421@main): Crash when restoring session state while a provisional history item is set https://bugs.webkit.org/show_bug.cgi?id=283924 rdar://140463108
Reviewed by Wenson Hsieh.

When session state is restored, the current history item index is updated, but the provisional index is
not cleared. So, if a provisional index was set before the session was restored, it will be stale and
point to an index that is incorrect or out-of-bounds `m_entries`.

* Source/WebKit/UIProcess/WebBackForwardList.cpp:
(WebKit::WebBackForwardList::restoreFromState):
* Tools/TestWebKitAPI/Tests/WebKit/WKBackForwardListTests.mm:
(TEST(WKBackForwardList, RestoreSessionStateResetProvisionalItem)):

Canonical link: https://commits.webkit.org/287248@main