SVGToOTFFontConversion.cpp:169:29: error: use of overloaded operator '[]' is ambiguous (with operand types 'ASCIILiteral' and 'int') https://bugs.webkit.org/show_bug.cgi?id=283768
Reviewed by Fujii Hironori.

Specify the type of the integer literal to be size_t to disambiguate
which operator[] to use. I'm not actually sure where the built-in
operator[]s are coming from, but that probably doesn't matter since we
surely want to use the one defined in ASCIILiteral.

* Source/WebCore/svg/SVGToOTFFontConversion.cpp:
(WebCore::SVGToOTFFontConverter::append32BitCode):
(WebCore::SVGToOTFFontConverter::appendTable):

Canonical link: https://commits.webkit.org/287287@main