REGRESSION(287315@main): css-rhythm/parsing/block-step-computed crashing with ASSERT_NOT_IMPLEMENTED_YET https://bugs.webkit.org/show_bug.cgi?id=284021 rdar://140893010
Reviewed by Tim Nguyen.

In 287315@main we did some refactoring and added some
ASSERT_NOT_IMPLEMENTED_YET()s on unsupported block-step-sizing content.
However, we failed to guard the call to performBlockStepSizing on
currently supported content so this results in block-step-computed.html
triggering the assert.

To fix this we can synchronize the logic used to guard the call
performBlockStepSizing and triggering the assert into a helper function.
We want to include a check to make sure block-step-insert is set to
margin-box since that is currently the only value we support for layout.

* Source/WebCore/rendering/RenderBlockFlow.cpp:
(WebCore::BlockStepSizing::childHasSupportedStyle):
(WebCore::RenderBlockFlow::performBlockStepSizing const):
(WebCore::RenderBlockFlow::layoutBlockChild):

Canonical link: https://commits.webkit.org/287347@main