Skip anchor-download-redirect-cross-origin.html on WK1 https://bugs.webkit.org/show_bug.cgi?id=284068 rdar://140939708
Reviewed by Alex Christensen.

The original fix is only intended for WK2 and there is not plan to add a similar fix for WK1.
The same applies to the test added to validate that fix.
Missing to add [ Skip ] for WK1 originally was an oversight. My bad.

* LayoutTests/platform/mac-wk1/TestExpectations:

Canonical link: https://commits.webkit.org/287418@main