Video loses audio after entering light spill rdar://140805541
Reviewed by Jer Noble.

In 286853@main we switched to use an AVSampleBufferVideoRenderer if available
and use it on a workqueue.
However, it has since appeared that its use over multiple thread is broken (rdar://139910776)

For now, if we do not use a decompression session, only ever use the
renderer given on construction (either an AVSampleBufferVideoRenderer if in
light spill, or an AVSampleBufferDisplayLayer any other time) on the main thread.

If a decompression session is set, then we will only ever accessed the AVSampleBufferVideoRenderer
on our queue.

Manually tested.

* Source/WebCore/platform/graphics/cocoa/VideoMediaSampleRenderer.mm:
(WebCore::VideoMediaSampleRenderer::~VideoMediaSampleRenderer):
(WebCore::VideoMediaSampleRenderer::maybeBecomeReadyForMoreMediaData):
(WebCore::VideoMediaSampleRenderer::stopRequestingMediaData):
(WebCore::VideoMediaSampleRenderer::resetReadyForMoreSample):

Canonical link: https://commits.webkit.org/287422@main