[AutoInstall] Remove any manifest item on failure https://bugs.webkit.org/show_bug.cgi?id=284086
Reviewed by Jonathan Bedard.

We had an EWS bot in a state where the manifest said something was
installed but it was not, likely due to installation failing.

When we're changing what version of a package is installed,
installation failing can lead us to a state where we have neither the
new nor old version installed, but the manifest would claim that the
old version still was.

To fix this, we delete any relevant manifest item on failure, which
will force a reinstall (of either version) in the future.

Untested, because of https://bugs.webkit.org/show_bug.cgi?id=284070.

* Tools/Scripts/libraries/webkitcorepy/webkitcorepy/autoinstall.py:
(Package.install):

Canonical link: https://commits.webkit.org/287426@main