Cmd+Shift+Up/Down and Cmd+Opt+Up/Down shouldn't scroll a web page https://bugs.webkit.org/show_bug.cgi?id=282852 rdar://139537723
Reviewed by Abrar Rahman Protyasha and Wenson Hsieh.

The designated shortcut for "Scroll to Top/Bottom" is Cmd+Up/Down Arrow.
Cmd+Shift/Opt/Ctrl+Up/Down are not valid key combinations for web page
scrolling, and might be designated shortcuts for other actions instead.

We already handled the case where Shift is pressed in 287161@main. This patch
ensures that while handling an Up/Down keyboard event with the Command key
pressed, we return a `Document` scroll granularity from `scrollGranularityForKeyboardEvent()`
only if no other modifier is pressed at the same time. If another modifier is
found to be pressed, we don't return a valid granularity so that no scroll
occurs.

Since `shouldBeTrue()`/`shouldBeEqual()` calls in the layout tests lead to debug
output that affects the scroll position, these need to be called with the
`quiet` parameter set to true.

* LayoutTests/fast/scrolling/ios/scroll-to-beginning-and-end-of-document-expected.txt:
* LayoutTests/fast/scrolling/ios/scroll-to-beginning-and-end-of-document.html:
* LayoutTests/fast/scrolling/mac/scroll-to-beginning-and-end-of-document-expected.txt: Copied from LayoutTests/fast/scrolling/ios/scroll-to-beginning-and-end-of-document-expected.txt.
* LayoutTests/fast/scrolling/mac/scroll-to-beginning-and-end-of-document.html: Added.
* Source/WebCore/platform/KeyboardScrollingAnimator.cpp:
(WebCore::scrollGranularityForKeyboardEvent):

Canonical link: https://commits.webkit.org/287437@main