Remove non-standard `webkitBackingStorePixelRatio`
https://bugs.webkit.org/show_bug.cgi?id=270057
rdar://problem/123980544

Reviewed by Tim Nguyen.

This patch is to align WebKit with Gecko / Firefox and Blink / Chromium.

`webkitBackingStorePixelRatio` was removed from Blink in 2014 with below
commit:

Commit: https://github.com/chromium/chromium/commit/07ee81b34997614fcc8785df2aefaa3dd2092306

It was updated to return value of '1' in 133750@main for high-dpi devices as well.
In other browser, it returns 'undefined' now, so this is to align with them.

* Source/WebCore/html/canvas/CanvasRenderingContext2D.idl:
* Source/WebCore/html/canvas/CanvasRenderingContext2DBase.h:

> Deleted Tests & Test Expectations:
* LayoutTests/fast/canvas/canvas-resize-reset-pixelRatio.html:
* LayoutTests/fast/canvas/canvas-resize-reset-pixelRatio-expected.txt:
* LayoutTests/fast/canvas/2d.backingStorePixelRatio.html:
* LayoutTests/fast/canvas/2d.backingStorePixelRatio-expected.txt:

Canonical link: https://commits.webkit.org/287523@main