macOS Ventura: dSYM build error because "an offset exceeds 4GB" rdar://140943450 https://bugs.webkit.org/show_bug.cgi?id=284194
Reviewed by Alexey Proskuryakov.

On the Ventura toolchain, a bug in dsymutil prevents constructing
universal DWARF binaries where the total size of the slices is over 4GB.
This is technically supported by the format, which merely requires that
a slice's *start address* is less than a 4GB offset into the file.

WebCore appears to have passed this mark recently. As a workaround,
build dSYMs for this configuration with line tables only.

* Source/WebCore/Configurations/WebCore.xcconfig:

Canonical link: https://commits.webkit.org/287557@main