REGRESSION (287477@main): [ iOS ] editing/selection/ios/tap-to-change-selection-after-accepting-inline-prediction.html is constantly timing out on iOS 18, consistently on 17. https://bugs.webkit.org/show_bug.cgi?id=284380 rdar://141228404
Reviewed by Abrar Rahman Protyasha.

Speculatively try to fix this flaky test by making the initial tap target taller, and replacing the
hard-coded tap location with a point computed from the bounds of the target element. Telemetry taken
from test runners shows that we're failing to even tap on the contentEditable body to focus it —
it's possible that this is because the hit-tested location is obscured by either the status bar, or
safe area inset.

* LayoutTests/editing/selection/ios/tap-to-change-selection-after-accepting-inline-prediction.html:

Canonical link: https://commits.webkit.org/287635@main