extractResolutionReport crash https://bugs.webkit.org/show_bug.cgi?id=284385 rdar://133489260
Reviewed by David Kilzer.

This is a speculative fix.

Add a retain to the resolution report to keep it live
long enough so that the dictionary construction gets a valid string from nw_xx apis

Also add a retain to the nw_path as it might also cause a future crash.

Also, add a null check for interfaces.get().

* Source/WebKit/NetworkProcess/cocoa/NetworkSessionCocoa.mm:
(extractResolutionReport):

Canonical link: https://commits.webkit.org/287651@main